refactor: begin refactoring into better layered architecture
