+++ backend/src/handlers/write.js
@@ -1,12 +1,12 @@
-const addCommentToVideo = require("../lib/add-comment-to-video");
-const addTag = require("../lib/add-tag");
-const addViewToVideo = require("../lib/add-view-to-video");
-const deleteTag = require("../lib/delete-tag");
+const addCommentToVideo = require("../application/add-comment-to-video");
+const addTag = require("../persistence/add-tag");
+const addViewToVideo = require("../application/add-view-to-video");
+const deleteTag = require("../persistence/delete-tag");
 const deleteVideo = require("../lib/delete-video");
 const ResponseFactory = require("../lib/factories/ResponseFactory");
-const getAuthToken = require("../lib/get-auth-token");
+const getAuthToken = require("../utility/get-auth-token");
 const isUserAnAdmin = require("../lib/is-user-an-admin");
-const updateVideo = require("../lib/update-video");
+const updateVideo = require("../persistence/update-video");
 
 module.exports.addViewToVideo = async event => {
 	try {
