+++ backend/src/lib/get-tags-available-to-user.js
@@ -1,4 +1,4 @@
-const getUser = require("./get-user");
+const getUser = require("../persistence/get-user");
 
 module.exports = async userHash => {
 	const user = await getUser(userHash);
