+++ backend/src/utility/get-auth-token.js
