+++ backend/src/utility/get-random-string.js
